Unified common mixin code
