+++ common/src/main/java/org/valkyrienskies/mod/mixin/mod_compat/create/accessors/Matrix3dAccessor.java
@@ -1,4 +1,4 @@
-package org.valkyrienskies.mod.forge.mixin.compat.create;
+package org.valkyrienskies.mod.mixin.mod_compat.create.accessors;
 
 import com.simibubi.create.foundation.collision.Matrix3d;
 import org.spongepowered.asm.mixin.Mixin;
