+++ fabric/src/main/resources/valkyrienskies-fabric.mixins.json
@@ -8,8 +8,6 @@
     "compat.cc_restitched.MixinTurtleBrain",
     "compat.cc_restitched.MixinTurtleMoveCommand",
     "compat.cc_restitched.MixinWirelessNetwork",
-    "compat.create.Matrix3dAccessor",
-    "compat.create.MixinAbstractContraptionEntity",
     "compat.create.MixinBlockBreakingKineticTileEntity",
     "compat.create.MixinBlocks",
     "compat.create.MixinControlledContraptionEntity",
